+++ src/index.ts
@@ -4,6 +4,8 @@ export type * from './types/HttpRequest';
 export type * from './types/FileUploadResponse';
 export type * from './types/StreamResponse';
 export type * from './types/StreamParser';
+export type * from './types/Interceptors';
+export type * from './types/Hooks';
 
 // main exports
 export { HttpClient } from './HttpClient';